refactor: consolidate authentication logic into dedicated handlers for dashboard and admin layouts
