+++ src/app/admin/layout.tsx
@@ -1,31 +1,13 @@
-import { redirect } from 'next/navigation'
 import { DashboardHeader } from '~/components/dashboard-header'
 import { AdminSidebar } from '~/components/sidebar/admin-sidebar'
-import { auth } from '~/lib/auth'
-import { db } from '~/server/db'
-import { getSelectedProfile } from '~/server/user/queries'
+import { handleAdminAuth } from '~/lib/auth-handlers'
 
 export default async function DashboardLayout({
 	children,
 }: {
 	children: React.ReactNode
 }) {
-	const { session, user } = await auth()
-	const selectedProfile = await getSelectedProfile()
-	const accounts = await db.query.userAccounts.findMany({
-		where: (accounts, { eq }) => eq(accounts.userId, session.user.id),
-	})
-
-	if (!user.admin) {
-		redirect('/unauthorized')
-	}
-
-	const userData = {
-		admin: user.admin,
-		name: session.user.name || 'User',
-		email: session.user.email || '',
-		avatar: session.user.image ?? undefined,
-	}
+	const { userData, accounts, selectedProfile } = await handleAdminAuth()
 
 	return (
 		<AdminSidebar
